requirements.txt
opencv-python
numpy
detectron2
ultralytics
Pillow
requests